refactor(extract): move metadata manager into separate file
