+++ sector-builder/src/constants.rs
@@ -1,4 +1,4 @@
-pub const NUM_SEAL_WORKERS: usize = 2;
+pub const NUM_WORKERS: usize = 2;
 
 pub const FATAL_NOSEND_TASK: &str = "[run_blocking] could not send";
 pub const FATAL_NORECV_TASK: &str = "[run_blocking] could not recv";
